refactor: Rename schemas module to validators for better semantic clarity
Rename sloth/schemas.py to sloth/validators.py to better reflect the module's
actual purpose and improve code organization semantic alignment.

### 🔧 Core Changes:

**File Operations:**
- Renamed sloth/schemas.py → sloth/validators.py (file rename, no content changes)
- Updated import statements across 8 files to use validators module

**Updated Import Statements in:**
- sloth/__init__.py: Main package imports
- sloth/common.py: Auto-detection imports and SchemaValidatorFactory
- sloth/loaders.py: Data loader imports
- sloth/pdbml_converter.py: XML converter imports
- tests/test_imports.py: Test imports and module structure tests
- tests/test_pdbml_conversion.py: Validation error imports and XML validator
- tests/test_relationship_resolution.py: Schema validator imports
- tests/test_schemas.py: Core schema validation test imports + comments

**Classes/Functions Affected:**
- SchemaValidator: Base abstract validator class
- JSONSchemaValidator: JSON schema validation
- XMLSchemaValidator: XML/XSD schema validation
- YAMLSchemaValidator: YAML schema validation
- CSVSchemaValidator: CSV schema validation
- SchemaValidatorFactory: Factory for creating validators
- ValidationError: Schema validation exception
- ValidationSeverity: Error severity enumeration

### 📂 Files Modified: 9 files
- 1 file renamed: schemas.py → validators.py
- 8 files updated: Import statement changes

### 🎯 Benefits:

1. **Semantic Clarity**: Module name now accurately reflects its validation purpose
2. **Better Organization**: Clearer distinction between schema definitions and validation logic
3. **Improved Readability**: Code intent is more obvious from module naming
4. **Consistency**: Aligns with the broader plugin/validation architecture

### ✅ Quality Assurance:

- Pure rename operation - no functional changes
- All imports updated consistently across codebase
- Test comments updated to reflect new module name
- Maintains full backwards compatibility through proper import updates

This refactoring improves code organization by using more descriptive and
semantically accurate module naming while preserving all existing
